Measure per-protocol quote latencies and request counts for both cached routes and non-cached routes (#323)
* Revert "Chore: measure per protocol version getRoutesThenQuotes and getQuotes latencies (#316)"

This reverts commit b8413a9d110e4216687902da963f530e7c31da8a.

* add both cached and non-cached per-protocol latencies and request counts

* address feedback
