+++ src/routers/alpha-router/quoters/base-quoter.ts
@@ -5,13 +5,7 @@ import { Pool } from '@uniswap/v3-sdk';
 import _ from 'lodash';
 
 import { ITokenListProvider, ITokenProvider, ITokenValidatorProvider, TokenValidationResult } from '../../../providers';
-import {
-  CurrencyAmount,
-  log,
-  metric,
-  MetricLoggerUnit,
-  poolToString
-} from '../../../util';
+import { CurrencyAmount, log, poolToString } from '../../../util';
 import { MixedRoute, V2Route, V3Route } from '../../router';
 import { AlphaRouterConfig } from '../alpha-router';
 import { RouteWithValidQuote } from '../entities/route-with-valid-quote';
@@ -32,7 +26,6 @@ export abstract class BaseQuoter<Route extends V2Route | V3Route | MixedRoute> {
   protected chainId: ChainId;
   protected blockedTokenListProvider?: ITokenListProvider;
   protected tokenValidatorProvider?: ITokenValidatorProvider;
-  protected abstract quoterVersion: string
 
   constructor(
     tokenProvider: ITokenProvider,
@@ -114,41 +107,19 @@ export abstract class BaseQuoter<Route extends V2Route | V3Route | MixedRoute> {
     gasModel?: IGasModel<RouteWithValidQuote>,
     gasPriceWei?: BigNumber
   ): Promise<GetQuotesResult> {
-    const beforeGetRoutesThenQuotes = Date.now();
-
     return this.getRoutes(tokenIn, tokenOut, tradeType, routingConfig)
-      .then((routesResult) => {
-          const beforeGetQuotes = Date.now();
-
-          return this.getQuotes(
-            routesResult.routes,
-            amounts,
-            percents,
-            quoteToken,
-            tradeType,
-            routingConfig,
-            routesResult.candidatePools,
-            gasModel,
-            gasPriceWei
-          ).then((quotesResult) => {
-            metric.putMetric(
-              `${this.quoterVersion}OverallGetQuotesLoad`,
-              Date.now() - beforeGetQuotes,
-              MetricLoggerUnit.Milliseconds
-            );
-
-            return quotesResult
-          })
-        }
-      ).then((quotesResult) => {
-          metric.putMetric(
-            `${this.quoterVersion}OverallGetRoutesThenQuotesLoad`,
-            Date.now() - beforeGetRoutesThenQuotes,
-            MetricLoggerUnit.Milliseconds
-          );
-
-          return quotesResult;
-        }
+      .then((routesResult) =>
+        this.getQuotes(
+          routesResult.routes,
+          amounts,
+          percents,
+          quoteToken,
+          tradeType,
+          routingConfig,
+          routesResult.candidatePools,
+          gasModel,
+          gasPriceWei
+        )
       );
   }
