+++ src/routers/alpha-router/quoters/v3-quoter.ts
@@ -26,7 +26,6 @@ export class V3Quoter extends BaseQuoter<V3Route> {
   protected v3SubgraphProvider: IV3SubgraphProvider;
   protected v3PoolProvider: IV3PoolProvider;
   protected onChainQuoteProvider: IOnChainQuoteProvider;
-  protected override quoterVersion = 'V3';
 
   constructor(
     v3SubgraphProvider: IV3SubgraphProvider,
